+++ src/main/java/br/com/phmiranda/comunidade/controller/WebController.java
@@ -5,7 +5,7 @@
  * Date: 05/05/2020
  */
 
-package br.com.alura.comunidade.controller;
+package br.com.phmiranda.comunidade.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
