+++ src/main/java/br/com/phmiranda/comunidade/domain/dto/TopicoDto.java
@@ -5,9 +5,9 @@
  * Date: 05/05/2020
  */
 
-package br.com.alura.comunidade.entity.dto;
+package br.com.phmiranda.comunidade.domain.dto;
 
-import br.com.alura.comunidade.entity.model.Topico;
+import br.com.phmiranda.comunidade.domain.model.Topico;
 
 import java.time.LocalDateTime;
 import java.util.List;
