+++ src/main/java/br/com/phmiranda/comunidade/domain/dto/TopicoDtoDetalhar.java
@@ -5,10 +5,10 @@
  * Date: 23/06/2020
  */
 
-package br.com.alura.comunidade.entity.dto;
+package br.com.phmiranda.comunidade.domain.dto;
 
-import br.com.alura.comunidade.entity.model.StatusTopico;
-import br.com.alura.comunidade.entity.model.Topico;
+import br.com.phmiranda.comunidade.domain.model.StatusTopico;
+import br.com.phmiranda.comunidade.domain.model.Topico;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
