+++ src/main/java/br/com/phmiranda/comunidade/domain/dto/TopicoDtoFormAtualizar.java
@@ -5,10 +5,10 @@
  * Date: 11/07/2020
  */
 
-package br.com.alura.comunidade.entity.dto;
+package br.com.phmiranda.comunidade.domain.dto;
 
-import br.com.alura.comunidade.entity.model.Topico;
-import br.com.alura.comunidade.repository.TopicoRepository;
+import br.com.phmiranda.comunidade.domain.model.Topico;
+import br.com.phmiranda.comunidade.repository.TopicoRepository;
 import org.hibernate.validator.constraints.Length;
 
 import javax.validation.constraints.NotEmpty;
