+++ src/main/java/br/com/phmiranda/comunidade/domain/model/Topico.java
@@ -5,27 +5,45 @@
  * Date: 23/12/2020
  */
 
-package br.com.alura.comunidade.entity.model;
+package br.com.phmiranda.comunidade.domain.model;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
-@Table(name = "respostas")
-public class Resposta {
+@Table(name = "topicos")
+public class Topico {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
+	private String titulo;
 	private String mensagem;
-
-	@ManyToOne
-	private Topico topico;
 	private LocalDateTime dataCriacao = LocalDateTime.now();
+
+	@Enumerated(EnumType.STRING)
+	private StatusTopico status = StatusTopico.NAO_RESPONDIDO;
 
 	@ManyToOne
 	private Usuario autor;
-	private Boolean solucao = false;
+
+	@ManyToOne
+	private Curso curso;
+
+	@OneToMany(mappedBy = "topico")
+	private List<Resposta> respostas = new ArrayList<>();
+
+	public Topico() {
+
+	}
+
+	public Topico(String titulo, String mensagem, Curso curso) {
+		this.titulo = titulo;
+		this.mensagem = mensagem;
+		this.curso = curso;
+	}
 
 	@Override
 	public int hashCode() {
@@ -43,7 +61,7 @@ public class Resposta {
 			return false;
 		if (getClass() != obj.getClass())
 			return false;
-		Resposta other = (Resposta) obj;
+		Topico other = (Topico) obj;
 		if (id == null) {
 			if (other.id != null)
 				return false;
@@ -61,20 +79,20 @@ public class Resposta {
 		this.id = id;
 	}
 
+	public String getTitulo() {
+		return titulo;
+	}
+
+	public void setTitulo(String titulo) {
+		this.titulo = titulo;
+	}
+
 	public String getMensagem() {
 		return mensagem;
 	}
 
 	public void setMensagem(String mensagem) {
 		this.mensagem = mensagem;
-	}
-
-	public Topico getTopico() {
-		return topico;
-	}
-
-	public void setTopico(Topico topico) {
-		this.topico = topico;
 	}
 
 	public LocalDateTime getDataCriacao() {
@@ -85,6 +103,14 @@ public class Resposta {
 		this.dataCriacao = dataCriacao;
 	}
 
+	public StatusTopico getStatus() {
+		return status;
+	}
+
+	public void setStatus(StatusTopico status) {
+		this.status = status;
+	}
+
 	public Usuario getAutor() {
 		return autor;
 	}
@@ -93,12 +119,20 @@ public class Resposta {
 		this.autor = autor;
 	}
 
-	public Boolean getSolucao() {
-		return solucao;
+	public Curso getCurso() {
+		return curso;
 	}
 
-	public void setSolucao(Boolean solucao) {
-		this.solucao = solucao;
+	public void setCurso(Curso curso) {
+		this.curso = curso;
+	}
+
+	public List<Resposta> getRespostas() {
+		return respostas;
+	}
+
+	public void setRespostas(List<Resposta> respostas) {
+		this.respostas = respostas;
 	}
 
 }
