+++ src/main/java/br/com/phmiranda/comunidade/repository/CursoRepository.java
@@ -5,9 +5,9 @@
  * Date: 05/05/2020
  */
 
-package br.com.alura.comunidade.repository;
+package br.com.phmiranda.comunidade.repository;
 
-import br.com.alura.comunidade.entity.model.Curso;
+import br.com.phmiranda.comunidade.domain.model.Curso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CursoRepository extends JpaRepository<Curso, Long> {
