+++ src/main/java/br/com/phmiranda/comunidade/service/TopicoService.java
@@ -5,7 +5,7 @@
  * Date: 29/04/2021
  */
 
-package br.com.alura.comunidade.service;
+package br.com.phmiranda.comunidade.service;
 
 public class TopicoService {
     // comment
