+++ src/main/resources/application.desenv.properties
@@ -1,15 +1,15 @@
 # comentário.
-server.servlet.context-path=/comunidade
+server.servlet.context-path=/api/comunidade
 server.port=9000
 
 # comentário.
 spring.datasource.driverClassName=com.mysql.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/db_comunidade?useTimezone=true&serverTimezone=UTC
+spring.datasource.url=jdbc:mysql://localhost:3306/db_comunidade
 spring.datasource.username=desenv
 spring.datasource.password=desenv
 
 # comentário.
-spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 spring.datasource.data=data.sql
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
